feat: refactored faucet (#1139)
* chore: savepoint

* chore: savepoint

* feat: testethiq faucet support
