+++ apps/shell/src/components/header-mobile.tsx
@@ -1,7 +1,7 @@
 'use client';
 import dynamic from 'next/dynamic';
 import { useAccount, useChains } from 'wagmi';
-import { haqqTestedge2 } from 'wagmi/chains';
+import { FAUCET_CHAINS } from '@haqq/shell-shared';
 import { HeaderMobile } from '@haqq/shell-ui-kit';
 import { useFilteredLinks } from '../hooks/use-filtered-header-links';
 import { useLocaleSwitcher } from '../hooks/use-locale-switcher';
@@ -16,7 +16,7 @@ const Web3ConnectButtonsMobile = dynamic(async () => {
 export function AppHeaderMobile({ className }: { className?: string }) {
   const chains = useChains();
   const { chain = chains[0] } = useAccount();
-  const isTestedge = chain.id === haqqTestedge2.id;
+  const isTestedge = FAUCET_CHAINS.includes(chain.id);
   const links = useFilteredLinks(chain);
   const { switchLocale, locales, currentLocale } = useLocaleSwitcher();
