+++ libs/shared/src/utils/chains.ts
@@ -2,6 +2,7 @@ import { haqqMainnet, haqqTestedge2, sepolia } from 'viem/chains';
 import { haqqTestethiq } from './bridge';
 
 export const bridgeSupportedChains = [haqqTestethiq, sepolia];
+export const faucetSupportedChains = [haqqTestedge2, haqqTestethiq];
 
 export const baseSupportedChains = [haqqMainnet, haqqTestedge2] as const;
